+++ src/extension.ts
@@ -58,6 +58,9 @@ export async function activate(context: vscode.ExtensionContext) {
 
 	// Register the mouse-over hover providers
 	await application.RegisterHoverProvidersAsync(context);
+
+	// Register code completion providers
+	await application.RegisterCompletionProvidersAsync(context);
 }
 
 // this method is called when your extension is deactivated